backend/requirements.txt
django==3.2.9
gunicorn==20.0.4
djangorestframework==3.12.4
pytest==6.2.5
pytest-django==4.4.0